+++ grunnur/__init__.py
@@ -1,4 +1,5 @@
 from . import dtypes, functions
+from .adapter_base import DeviceParameters
 from .api import (
     API,
     all_api_ids,
@@ -6,9 +7,9 @@ from .api import (
     opencl_api_id,
 )
 from .array import Array, ArrayLike, MultiArray
-from .array_metadata import ArrayMetadata, ArrayMetadataLike
+from .array_metadata import ArrayMetadata, AsArrayMetadata
 from .buffer import Buffer
-from .context import Context
+from .context import BoundDevice, Context
 from .device import Device, DeviceFilter
 from .device_discovery import (
     platforms_and_devices_by_mask,